8.1.4 refactor main.js, remove mock data.
