+++ src/view/event-edit.js
@@ -4,8 +4,7 @@ import {createEventTypeTemplate} from "./event-type-template.js";
 import {createEventDestinationTemplate} from "./event-destination-template.js";
 import {createEventEditOffersTemplate} from "./event-offers-template.js";
 import {createEventEditDescriptionTemplate} from "./event-description-template.js";
-import {getTimes} from "../utils/event.js";
-import {getEventOffers, getNewInformation, ALL_OFFERS} from "../mock/event.js";
+import {getTimes, getEventOffers, getNewInformation, ALL_OFFERS} from "../utils/event.js";
 import dayjs from "dayjs";
 import flatpickr from "flatpickr";
